+++ Tests/MapboxCoreNavigationTests/RouteControllerTests.swift
@@ -15,7 +15,7 @@ class RouteControllerTests: TestCase {
 
     override func tearDown() {
         replayManager = nil
-        NavigationRouter.__testRoutesStub = nil
+        MapboxRoutingProvider.__testRoutesStub = nil
         super.tearDown()
     }
     
@@ -29,7 +29,7 @@ class RouteControllerTests: TestCase {
         let locationManager = ReplayLocationManager(locations: locations)
         replayManager = locationManager
         let equivalentRouteOptions = NavigationRouteOptions(navigationMatchOptions: options)
-        let routeController = RouteController(alongRouteAtIndex: 0, in: routeResponse, options: equivalentRouteOptions, routingProvider: NavigationRouter(.offline), dataSource: self)
+        let routeController = RouteController(alongRouteAtIndex: 0, in: routeResponse, options: equivalentRouteOptions, routingProvider: MapboxRoutingProvider(.offline), dataSource: self)
         locationManager.delegate = routeController
         let routerDelegateSpy = RouterDelegateSpy()
         routeController.delegate = routerDelegateSpy
@@ -75,7 +75,7 @@ class RouteControllerTests: TestCase {
         let routeController = RouteController(alongRouteAtIndex: 0,
                                               in: routeResponse,
                                               options: navOptions,
-                                              routingProvider: NavigationRouter(.offline),
+                                              routingProvider: MapboxRoutingProvider(.offline),
                                               dataSource: self)
 
         let routerDelegateSpy = RouterDelegateSpy()
@@ -111,7 +111,7 @@ class RouteControllerTests: TestCase {
             didRerouteCalled.fulfill()
         }
         
-        NavigationRouter.__testRoutesStub = { (options, completionHandler) in
+        MapboxRoutingProvider.__testRoutesStub = { (options, completionHandler) in
             DispatchQueue.main.async {
                 completionHandler(Directions.Session(options, .mocked),
                                   .success(routeResponse))